refactor: split into hook
