+++ apps/web/src/app/photo/[photoId]/page.tsx
@@ -1,24 +1,13 @@
 'use client';
 
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useParams, useRouter, useSearchParams } from 'next/navigation';
-import {
-  getGetPhotosQueryKey,
-  getGetSelectableAlbumsQueryKey,
-  useGetPhotoDetail,
-  useGetPhotos,
-  useGetSelectableAlbums,
-} from '@repo/api-client';
 import AlbumIcon from '@/assets/images/album.svg';
 import CommentIcon from '@/assets/images/comment.svg';
 import DateIcon from '@/assets/images/date.svg';
 import Chip from '@/components/buttons/chip/Chip';
 import MenuHeader from '@/components/header/menu/MenuHeader';
+import { useLongPress, usePhotoData, usePhotoSlider } from './_hooks';
 import * as S from './page.styles';
-
-const LONG_PRESS_DURATION = 1500;
-// TODO: 사용자 컨텍스트에서 가져오도록 수정
-const TEMP_USER_ID = 1;
 
 export default function PhotoViewPage() {
   const router = useRouter();
@@ -30,117 +19,27 @@ export default function PhotoViewPage() {
     ? Number(searchParams.get('albumId'))
     : undefined;
 
-  const [isOverlayVisible, setIsOverlayVisible] = useState(true);
-  const [currentPhotoIndex, setCurrentPhotoIndex] = useState(0);
-  const longPressTimerRef = useRef<NodeJS.Timeout | null>(null);
-  const thumbnailContainerRef = useRef<HTMLDivElement>(null);
-
-  const { data: photoDetail, isLoading: isPhotoLoading } = useGetPhotoDetail(photoId);
-
-  // 앨범 목록 조회 (albumId가 없을 때 albumName으로 찾기 위함)
-  const selectableAlbumsParams = { userId: TEMP_USER_ID };
-  const { data: selectableAlbums } = useGetSelectableAlbums(selectableAlbumsParams, {
-    query: {
-      queryKey: getGetSelectableAlbumsQueryKey(selectableAlbumsParams),
-      enabled: !albumIdFromQuery && !!photoDetail?.albumName,
-    },
+  const { photoDetail, photos, isPhotoLoading } = usePhotoData({
+    photoId,
+    albumIdFromQuery,
   });
 
-  // albumId 결정: 쿼리 파라미터 우선, 없으면 albumName으로 매칭
-  const albumId = useMemo(() => {
-    if (albumIdFromQuery) return albumIdFromQuery;
+  const {
+    currentPhotoIndex,
+    currentPhoto,
+    thumbnailContainerRef,
+    handlePrevPhoto,
+    handleNextPhoto,
+    handleThumbnailClick,
+  } = usePhotoSlider({ photos, initialPhotoId: photoId });
 
-    if (selectableAlbums?.albums && photoDetail?.albumName) {
-      const matchedAlbum = selectableAlbums.albums.find(
-        (album) => album.title === photoDetail.albumName
-      );
-      return matchedAlbum?.id;
-    }
-
-    return undefined;
-  }, [albumIdFromQuery, selectableAlbums?.albums, photoDetail?.albumName]);
-
-  const albumParams = { albumId: albumId ?? 0 };
-  const { data: albumPhotos } = useGetPhotos(albumParams, {
-    query: {
-      queryKey: getGetPhotosQueryKey(albumParams),
-      enabled: !!albumId,
-    },
-  });
-
-  const photos = useMemo(() => {
-    if (!albumPhotos?.albums?.[0]?.photos) return [];
-    return albumPhotos.albums[0].photos;
-  }, [albumPhotos]);
-
-  const currentPhoto = useMemo(() => {
-    if (photos.length > 0) {
-      return photos[currentPhotoIndex];
-    }
-    return null;
-  }, [photos, currentPhotoIndex]);
+  const { isOverlayVisible, longPressHandlers } = useLongPress();
 
   const displayPhotoUrl = currentPhoto?.url || photoDetail?.url;
 
-  useEffect(() => {
-    if (photos.length > 0 && photoId) {
-      const index = photos.findIndex((p) => p.id === photoId);
-      if (index !== -1) {
-        setCurrentPhotoIndex(index);
-      }
-    }
-  }, [photos, photoId]);
-
-  useEffect(() => {
-    if (thumbnailContainerRef.current && photos.length > 0) {
-      const container = thumbnailContainerRef.current;
-      const thumbnailWidth = 40;
-      const activeThumbnailWidth = 56;
-      const gap = 8;
-      const containerWidth = container.offsetWidth;
-      const scrollPosition =
-        currentPhotoIndex * (thumbnailWidth + gap) -
-        containerWidth / 2 +
-        activeThumbnailWidth / 2;
-      container.scrollTo({ left: scrollPosition, behavior: 'smooth' });
-    }
-  }, [currentPhotoIndex, photos.length]);
-
-  const handleBack = useCallback(() => {
+  const handleBack = () => {
     router.back();
-  }, [router]);
-
-  const handleLeftAreaClick = useCallback(() => {
-    if (photos.length > 0 && currentPhotoIndex > 0) {
-      setCurrentPhotoIndex((prev) => prev - 1);
-    }
-  }, [photos.length, currentPhotoIndex]);
-
-  const handleRightAreaClick = useCallback(() => {
-    if (photos.length > 0 && currentPhotoIndex < photos.length - 1) {
-      setCurrentPhotoIndex((prev) => prev + 1);
-    }
-  }, [photos.length, currentPhotoIndex]);
-
-  const handleLongPressStart = useCallback(() => {
-    longPressTimerRef.current = setTimeout(() => {
-      setIsOverlayVisible(false);
-    }, LONG_PRESS_DURATION);
-  }, []);
-
-  const handleLongPressEnd = useCallback(() => {
-    if (longPressTimerRef.current) {
-      clearTimeout(longPressTimerRef.current);
-      longPressTimerRef.current = null;
-    }
-    longPressTimerRef.current = setTimeout(() => {
-      setIsOverlayVisible(true);
-    }, LONG_PRESS_DURATION);
-  }, []);
-
-  const handleThumbnailClick = useCallback((index: number) => {
-    setCurrentPhotoIndex(index);
-  }, []);
+  };
 
   if (isPhotoLoading) {
     return (
@@ -159,17 +58,11 @@ export default function PhotoViewPage() {
   }
 
   return (
-    <S.Container
-      onMouseDown={handleLongPressStart}
-      onMouseUp={handleLongPressEnd}
-      onMouseLeave={handleLongPressEnd}
-      onTouchStart={handleLongPressStart}
-      onTouchEnd={handleLongPressEnd}
-    >
+    <S.Container {...longPressHandlers}>
       <S.PhotoBackground $url={displayPhotoUrl || ''} />
 
-      <S.TouchAreaLeft onClick={handleLeftAreaClick} />
-      <S.TouchAreaRight onClick={handleRightAreaClick} />
+      <S.TouchAreaLeft onClick={handlePrevPhoto} />
+      <S.TouchAreaRight onClick={handleNextPhoto} />
 
       {isOverlayVisible && (
         <>
